+++ packages/stargate/src/queryclient/queryclient.ts
@@ -1,23 +1,11 @@
 /* eslint-disable no-dupe-class-members, @typescript-eslint/ban-types, @typescript-eslint/naming-convention */
-import { iavlSpec, ics23, tendermintSpec, verifyExistence, verifyNonExistence } from "@confio/ics23";
-import { toAscii, toHex } from "@cosmjs/encoding";
 import { firstEvent } from "@cosmjs/stream";
 import { CometClient, tendermint34 } from "@cosmjs/tendermint-rpc";
-import { arrayContentEquals, assert, assertDefined, isNonNullObject, sleep } from "@cosmjs/utils";
+import { assert, assertDefined, isNonNullObject, sleep } from "@cosmjs/utils";
 import { ProofOps } from "cosmjs-types/tendermint/crypto/proof";
 import { Stream } from "xstream";
 
 type QueryExtensionSetup<P> = (base: QueryClient) => P;
-
-function checkAndParseOp(op: tendermint34.ProofOp, kind: string, key: Uint8Array): ics23.CommitmentProof {
-  if (op.type !== kind) {
-    throw new Error(`Op expected to be ${kind}, got "${op.type}`);
-  }
-  if (!arrayContentEquals(key, op.key)) {
-    throw new Error(`Proven key different than queried key.\nQuery: ${toHex(key)}\nProven: ${toHex(op.key)}`);
-  }
-  return ics23.CommitmentProof.decode(op.data);
-}
 
 export interface ProvenQuery {
   readonly key: Uint8Array;
@@ -510,90 +498,6 @@ export class QueryClient {
 
   public constructor(cometClient: CometClient) {
     this.cometClient = cometClient;
-  }
-
-  /**
-   * Queries the database store with a proof, which is then verified.
-   *
-   * Please note: the current implementation trusts block headers it gets from the PRC endpoint.
-   */
-  public async queryStoreVerified(
-    store: string,
-    queryKey: Uint8Array,
-    desiredHeight?: number,
-  ): Promise<QueryStoreResponse> {
-    const { height, proof, key, value } = await this.queryRawProof(store, queryKey, desiredHeight);
-
-    const subProof = checkAndParseOp(proof.ops[0], "ics23:iavl", queryKey);
-    const storeProof = checkAndParseOp(proof.ops[1], "ics23:simple", toAscii(store));
-
-    // this must always be existence, if the store is not a typo
-    assert(storeProof.exist);
-    assert(storeProof.exist.value);
-
-    // this may be exist or non-exist, depends on response
-    if (!value || value.length === 0) {
-      // non-existence check
-      assert(subProof.nonexist);
-      // the subproof must map the desired key to the "value" of the storeProof
-      verifyNonExistence(subProof.nonexist, iavlSpec, storeProof.exist.value, queryKey);
-    } else {
-      // existence check
-      assert(subProof.exist);
-      assert(subProof.exist.value);
-      // the subproof must map the desired key to the "value" of the storeProof
-      verifyExistence(subProof.exist, iavlSpec, storeProof.exist.value, queryKey, value);
-    }
-
-    // the store proof must map its declared value (root of subProof) to the appHash of the next block
-    const header = await this.getNextHeader(height);
-    verifyExistence(storeProof.exist, tendermintSpec, header.appHash, toAscii(store), storeProof.exist.value);
-
-    return { key, value, height };
-  }
-
-  public async queryRawProof(
-    store: string,
-    queryKey: Uint8Array,
-    desiredHeight?: number,
-  ): Promise<ProvenQuery> {
-    const { key, value, height, proof, code, log } = await this.cometClient.abciQuery({
-      // we need the StoreKey for the module, not the module name
-      // https://github.com/cosmos/cosmos-sdk/blob/8cab43c8120fec5200c3459cbf4a92017bb6f287/x/auth/types/keys.go#L12
-      path: `/store/${store}/key`,
-      data: queryKey,
-      prove: true,
-      height: desiredHeight,
-    });
-
-    if (code) {
-      throw new Error(`Query failed with (${code}): ${log}`);
-    }
-
-    if (!arrayContentEquals(queryKey, key)) {
-      throw new Error(`Response key ${toHex(key)} doesn't match query key ${toHex(queryKey)}`);
-    }
-
-    if (!height) {
-      throw new Error("No query height returned");
-    }
-    if (!proof || proof.ops.length !== 2) {
-      throw new Error(`Expected 2 proof ops, got ${proof?.ops.length ?? 0}. Are you using stargate?`);
-    }
-
-    // we don't need the results, but we can ensure the data is the proper format
-    checkAndParseOp(proof.ops[0], "ics23:iavl", key);
-    checkAndParseOp(proof.ops[1], "ics23:simple", toAscii(store));
-
-    return {
-      key: key,
-      value: value,
-      height: height,
-      // need to clone this: readonly input / writeable output
-      proof: {
-        ops: [...proof.ops],
-      },
-    };
   }
 
   /**